Move direct archive downloader into core since it's not exclusive to CDA
