+++ speasy/core/direct_archive_downloader/direct_archive_downloader.py
